Comentarios inseridos e separação de componentes
