--- desafio_estagio/app/Http/Controllers/ProjectController.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class ProjectController extends Controller
-{
-    public function index()
-    {
-        return response()->json(['message' => 'List of projects']);
-    }
-}
