+++ desafio_estagio/resources/css/HeroSection.css
@@ -1,0 +1,34 @@
+
+
+.hero-title {
+    margin: 0 0 20px 0;
+    font-size: 48px;
+    font-weight: 700;
+    line-height: 56px;
+    color: #ffffff;
+}
+
+.hero-subtitle {
+    color: rgba(255, 255, 255, 0.8);
+    margin-bottom: 40px;
+    font-size: 24px;
+}
+
+
+
+.hero-img {
+    width: 100%;
+    max-width: 100%;
+    height: auto;
+    margin-right: -30px;
+    border-radius: 20px;
+}
+
+.discover-artists {
+    text-align: center;
+    margin-top: 40px;
+}
+
+.artist-cards {
+    margin: 0 -0.5rem;
+}
